cnn_benchmark/requirements.txt
boto3==1.9.243
grpcio==1.16.1
h5py==2.9.0
hdf5==1.10.4
imageio==2.6.0
jsonnet==0.14.0
matplotlib==3.1.1
mlflow==1.3.0
numpy==1.16.4
opencv-python==4.1.1.26
overrides
pandas==0.25.0
pillow==6.2.0
protobuf==3.9.2
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.1.0
tensorflow-gpu==1.10.0
tensorpack==0.9.8
tqdm
urllib3=1.24.3